Updating the decorators based on new model. I could not get the inheritance to work properly when subclassing the oauth decorator. Perhaps it is something we can look into further at some point, but I've realize we probably don't want to mess with that too much as getting it wrong could have disastrous effects anyway. This still combines the previous application and domain admin decorators into one new one, leaving the application admin as a standalone as necessary and removing the domain standalone. This also renames dasher to domain as appropriate.
