+++ user.py
@@ -3,9 +3,7 @@
 import admin
 from appengine_config import JINJA_ENVIRONMENT
 from ast import literal_eval
-from auth import OAUTH_DECORATOR
 import base64
-import dasher_admin
 from datastore import DomainVerification
 from datastore import ProxyServer
 from datastore import User
@@ -186,8 +184,8 @@ class ListUsersHandler(webapp2.RequestHandler):
 
   # pylint: disable=too-few-public-methods
 
-  @OAUTH_DECORATOR.oauth_required
-  @dasher_admin.DasherAdminAuthRequired
+  @admin.OAUTH_USER_SCOPE_DECORATOR.oauth_required
+  @admin.RequireAppAndDomainAdmin(admin.OAUTH_USER_SCOPE_DECORATOR)
   def get(self):
     """Output a list of all current users along with some metadata."""
     self.response.write(_RenderUserListTemplate())
@@ -199,8 +197,8 @@ class DeleteUserHandler(webapp2.RequestHandler):
 
   # pylint: disable=too-few-public-methods
 
-  @OAUTH_DECORATOR.oauth_required
-  @dasher_admin.DasherAdminAuthRequired
+  @admin.OAUTH_USER_SCOPE_DECORATOR.oauth_required
+  @admin.RequireAppAndDomainAdmin(admin.OAUTH_USER_SCOPE_DECORATOR)
   def get(self):
     """Delete the user corresponding to the passed in key.
 
@@ -217,8 +215,8 @@ class ListTokensHandler(webapp2.RequestHandler):
 
   # pylint: disable=too-few-public-methods
 
-  @OAUTH_DECORATOR.oauth_required
-  @dasher_admin.DasherAdminAuthRequired
+  @admin.OAUTH_USER_SCOPE_DECORATOR.oauth_required
+  @admin.RequireAppAndDomainAdmin(admin.OAUTH_USER_SCOPE_DECORATOR)
   def get(self):
     """Output a list of all current users along with each's token."""
     self.response.write(_RenderTokenListTemplate())
@@ -230,8 +228,8 @@ class GetInviteCodeHandler(webapp2.RequestHandler):
 
   # pylint: disable=too-few-public-methods
 
-  @OAUTH_DECORATOR.oauth_required
-  @dasher_admin.DasherAdminAuthRequired
+  @admin.OAUTH_USER_SCOPE_DECORATOR.oauth_required
+  @admin.RequireAppAndDomainAdmin(admin.OAUTH_USER_SCOPE_DECORATOR)
   def get(self):
     """Output a list of all current users along with the requested token."""
     urlsafe_key = self.request.get('key')
@@ -247,8 +245,8 @@ class GetNewTokenHandler(webapp2.RequestHandler):
 
   # pylint: disable=too-few-public-methods
 
-  @OAUTH_DECORATOR.oauth_required
-  @dasher_admin.DasherAdminAuthRequired
+  @admin.OAUTH_USER_SCOPE_DECORATOR.oauth_required
+  @admin.RequireAppAndDomainAdmin(admin.OAUTH_USER_SCOPE_DECORATOR)
   def get(self):
     """Find the user matching the specified key and generate a new token."""
     urlsafe_key = self.request.get('key')
@@ -261,9 +259,8 @@ class AddUsersHandler(webapp2.RequestHandler):
 
   """Add users into the datastore."""
 
-  @admin.RequireAdmin
-  @OAUTH_DECORATOR.oauth_required
-  @dasher_admin.DasherAdminAuthRequired
+  @admin.OAUTH_ALL_SCOPES_DECORATOR.oauth_required
+  @admin.RequireAppAndDomainAdmin(admin.OAUTH_ALL_SCOPES_DECORATOR)
   def get(self):
     """Get the form for adding new users.
 
@@ -278,15 +275,18 @@ class AddUsersHandler(webapp2.RequestHandler):
     user_key = self.request.get('user_key')
     try:
       if get_all:
-        directory_service = GoogleDirectoryService(OAUTH_DECORATOR)
+        directory_service = GoogleDirectoryService(
+          admin.OAUTH_ALL_SCOPES_DECORATOR)
         directory_users = directory_service.GetUsers()
         self.response.write(_RenderAddUsersTemplate(directory_users))
       elif group_key is not None and group_key is not '':
-        directory_service = GoogleDirectoryService(OAUTH_DECORATOR)
+        directory_service = GoogleDirectoryService(
+          admin.OAUTH_ALL_SCOPES_DECORATOR)
         directory_users = directory_service.GetUsersByGroupKey(group_key)
         self.response.write(_RenderAddUsersTemplate(directory_users))
       elif user_key is not None and user_key is not '':
-        directory_service = GoogleDirectoryService(OAUTH_DECORATOR)
+        directory_service = GoogleDirectoryService(
+          admin.OAUTH_ALL_SCOPES_DECORATOR)
         directory_users = directory_service.GetUserAsList(user_key)
         self.response.write(_RenderAddUsersTemplate(directory_users))
       else:
@@ -294,10 +294,9 @@ class AddUsersHandler(webapp2.RequestHandler):
     except errors.HttpError as error:
       self.response.write(_RenderAddUsersTemplate([], error))
 
-  @admin.RequireAdmin
+  @admin.OAUTH_ALL_SCOPES_DECORATOR.oauth_required
+  @admin.RequireAppAndDomainAdmin(admin.OAUTH_ALL_SCOPES_DECORATOR)
   @xsrf.XSRFProtect
-  @OAUTH_DECORATOR.oauth_required
-  @dasher_admin.DasherAdminAuthRequired
   def post(self):
     """Add all of the selected users into the datastore."""
     users = self.request.get_all('selected_user')
@@ -317,7 +316,10 @@ APP = webapp2.WSGIApplication([
     ('/user/getInviteCode', GetInviteCodeHandler),
     ('/user/getNewToken', GetNewTokenHandler),
     ('/user/add', AddUsersHandler),
-    (OAUTH_DECORATOR.callback_path, OAUTH_DECORATOR.callback_handler()),
+    (admin.OAUTH_ALL_SCOPES_DECORATOR.callback_path,
+      admin.OAUTH_ALL_SCOPES_DECORATOR.callback_handler()),
+    (admin.OAUTH_USER_SCOPE_DECORATOR.callback_path,
+      admin.OAUTH_USER_SCOPE_DECORATOR.callback_handler()),
 ], debug=True)
 
 # This is the only way to catch exceptions from the oauth decorators.
